refactor(wgpu): 力度块专用 16B 字段 + ghost f64→u32 + tension f32
- ghost_notes 类型从 f64 改为 u32，消除 pencil 计算 u32→f64→GPU 回转 u32 的多余转换
- tension 从 i8 改为 f32 全链路统一，消除 /127 归一化转换
- 新建 VelocityBarInstance (16B = vec4<u32>)：tick/length/track/velocity 语义数据
- 新建 velocity_pipeline + vs_main_velocity：GPU 从 uniforms + track_colors 计算位置和颜色
- 三种力度显示模式统一为边框模式（fill + border），复用 fs_main
- 删除旧的遮挡排序（duration/start_tick/vel），改为 (tick, packed) 确定性排序
- automation/prepare.rs 新增 upload_track_colors，让 velocity shader 能查 tc[track]
- Uniforms 新增 value_zoom/value_scroll 字段，供 velocity shader 计算 Y 坐标
- 清理 velocity_display_mode 全链路：AudioSettings/UI/prepare/automation_panel/piano_view/layout

+++ crates/yinhe-egui/src/piano_view/drag.rs
@@ -156,7 +156,7 @@ pub(crate) fn sel_drag_frame(
     _track_colors: &[[f32; 3]],
     track_visible: &[bool],
     track_selected: &std::collections::HashSet<u16>,
-) -> (Vec<(f64, f64, u8, u16)>, Vec<(u16, u32, u8)>) {
+) -> (Vec<(u32, u32, u8, u16)>, Vec<(u16, u32, u8)>) {
     let note_drag_id = ui.id().with("note_drag_origin");
     let mut note_drag_origin: Option<(f64, f64)> =
         ui.data_mut(|d| d.get_persisted(note_drag_id)).unwrap_or(None);
@@ -240,7 +240,7 @@ pub(crate) fn sel_drag_frame(
     }
 
     // Note drag: use pre-computed data for ghost/hidden, store delta only on release
-    let mut ghost_notes: Vec<(f64, f64, u8, u16)> = Vec::new();
+    let mut ghost_notes: Vec<(u32, u32, u8, u16)> = Vec::new();
     let mut hidden_notes: Vec<(u16, u32, u8)> = Vec::new();
     if let Some((origin_tick, origin_key)) = note_drag_origin {
         if let Some(ref notes) = drag_notes {
@@ -259,7 +259,7 @@ pub(crate) fn sel_drag_frame(
                         let new_tick = (info.start_tick as i64 + dt).max(0) as u32;
                         let new_key = ((info.key as i32) + dk).clamp(0, 127) as u8;
                         let length = info.end_tick - info.start_tick;
-                        ghost_notes.push((new_tick as f64, (new_tick + length) as f64, new_key, info.track));
+                        ghost_notes.push((new_tick, new_tick + length, new_key, info.track));
                         hidden_notes.push((info.track, info.start_tick, info.key));
                     }
 
@@ -285,7 +285,7 @@ pub(crate) fn sel_drag_frame(
                         let new_tick = (info.start_tick as i64 + dt).max(0) as u32;
                         let new_key = ((info.key as i32) + dk).clamp(0, 127) as u8;
                         let length = info.end_tick - info.start_tick;
-                        ghost_notes.push((new_tick as f64, (new_tick + length) as f64, new_key, info.track));
+                        ghost_notes.push((new_tick, new_tick + length, new_key, info.track));
                         hidden_notes.push((info.track, info.start_tick, info.key));
                     }
                 }
